requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bitsandbytes==0.47.0
blessed==1.22.0
certifi==2025.8.3
charset-normalizer==3.4.3
datasets==4.1.1
deepspeed==0.17.6
dill==0.4.0
distro==1.9.0
einops==0.8.1
filelock==3.19.1
flash-attention==1.0.0
flash_attn==2.8.3
frozenlist==1.7.0
fsspec==2025.9.0
gpustat==1.1.1
h11==0.16.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
ninja==1.13.0
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.108.1
packaging==25.0
pandas==2.3.2
peft==0.17.1
propcache==0.3.2
psutil==7.1.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
trl==0.23.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
